CP-18618: If live patching fails, warn the user
Intermediate checkin. Currently has fake functionality, will always claim every involved host requires a reboot. If one host, uses ThreeButtonDialog to offer to reboot it, else uses a new RebootHostsDialog to let the user choose which, if any, to reboot. Uses the progress bar + log already on the page, which is probably the wrong design but works for now.

Signed-off-by: Callum McIntyre <callumiandavid.mcintyre@citrix.com>