CP-18618: Only advise we need a restart if this patch failed
Using patches_requiring_reboot.any() means that we advise that a live patch failed if the machine already needed rebooting for some other patch, which is incorrect. We now only advise that live patching failed if we have a reboot required for the specific patch bering run, which should have live patched.

Signed-off-by: Callum McIntyre <callumiandavid.mcintyre@citrix.com>